--- requirements.txt
@@ -1,22 +1,0 @@
-annotated-doc==0.0.4
-annotated-types==0.7.0
-anyio==4.13.0
-asyncpg==0.31.0
-click==8.3.2
-fastapi==0.135.3
-greenlet==3.2.4
-h11==0.16.0
-idna==3.11
-numpy==2.4.4
-pgvector==0.4.2
-psycopg==3.3.3
-psycopg-binary==3.3.3
-pydantic==2.12.5
-pydantic-settings==2.10.1
-pydantic_core==2.41.5
-python-dotenv==1.2.2
-SQLAlchemy==2.0.49
-starlette==1.0.0
-typing-inspection==0.4.2
-typing_extensions==4.15.0
-uvicorn==0.44.0
